+++ CHANGELOG.md
@@ -27,6 +27,7 @@
   - GET /roles/:id
   - GET /roles
   - DELETE /roles/:id
+- added new property to User model, user needs the role to create new user
 
 ## Removed